requirements.txt
absl_py==1.3.0
numpy==1.23.5
imageio==2.23.0
matplotlib==3.6.2
opencv_python==4.6.0.66
pyrender==0.1.45
tqdm==4.64.1
trimesh==3.9.43
networkx==2.5
warp_lang==0.7.2
tensorboard==2.13.0
imageio[ffmpeg]
imageio[pyav]
dqtorch @ git+https://github.com/MightyChaos/dqtorch